requirements.txt
numpy==2.1.3
opencv-python==4.10.0.84
pandas==2.2.3
pillow==11.0.0
